requirements.txt
Pillow==8.2.0
streamlit==0.82.0
